💥 Seperate service logic
