+++ app/services/user_service.py
@@ -84,44 +84,3 @@ class UserService:
         if not user:
             raise ValueError(f'User with id {user_id} not found')
         return self._serialize_user(user)
-
-    # TODO(GideokKim): User DAO가 필요 없으므로 나중에 새로 service 만들어서 관리해야 함
-    def update_user_memory_with_llm(self, user_id: str, summary: Optional[str], messages: List[BaseMessage]) -> Optional[str]:
-        """
-        summary와 messages, 기존 user_memory를 LLM에 입력해 장기기억(user_memory)을 통합/업데이트
-        """
-        prev_memory = self.get_user_memory(user_id) or ""
-
-        user_prompt = (
-        f"[기존 장기기억]\n{prev_memory}\n\n"
-        f"[최근 대화 요약]\n{summary or ''}\n\n"
-        f"[최근 메시지]\n{messages}"
-        )
-
-        llm_messages = [
-            {"role": "system", "content": USER_MEMORY_SYSTEM_PROMPT},
-            {"role": "user", "content": user_prompt}
-        ]
-        updated_memory = get_completion(llm_messages)
-        return self.update_user_memory(user_id, memory_data=updated_memory)
-
-    # TODO(GideokKim): User DAO가 필요 없으므로 나중에 새로 service 만들어서 관리해야 함
-    def initialize_agent_state(self, user_id: str) -> dict:
-        """대화 세션 시작 시 AgentState에 user_memory를 반영"""
-        user_memory = self.get_user_memory(user_id)
-        return {
-            "messages": [],
-            "summary": None,
-            "user_memory": user_memory,
-            "user_id": user_id,
-            "current_input": "",
-            "scratchpad": [],
-            "next_step": None
-        }
-
-    # TODO(GideokKim): User DAO가 필요 없으므로 나중에 새로 service 만들어서 관리해야 함
-    def save_user_memory_from_state(self, user_id: str, agent_state: dict) -> Optional[str]:
-        """대화 세션 종료 시 AgentState의 summary/messages를 활용해 user_memory를 LLM으로 업데이트"""
-        summary = agent_state.get("summary")
-        messages = agent_state.get("messages", [])
-        return self.update_user_memory_with_llm(user_id, summary, messages)
